requirements.txt
aiogram==3.2.0
aiohttp==3.9.1
PyYAML==6.0.1